Mudando de Double para BigDecimal os valor de custeio

+++ src/main/java/com/malbano/rural/model/dto/AcumuloDTO.java
@@ -17,7 +17,7 @@ public class AcumuloDTO {
     private String nomeRegiao;
     private String nomeUF;
     private String anoEmissao;
-    private Double vlCusteioTotal;
+    private BigDecimal vlCusteioTotal;
 
 
     public static AcumuloDTO create(DadosEntity dados) {
